Conflicts:
	27GP3.sdf
	27GP3.v12.suo
	Debug/27Gp3Shoot.exe
	Debug/27Gp3Shoot.ilk
	Debug/27Gp3Shoot.pdb

+++ 27Gp3Shoot/src/effect3D.cpp
@@ -77,7 +77,7 @@ namespace effect3D
   {
     //Ç®Ç®ÇÊÇª0.5secÇ≈è¡ñ≈
     alpha_ -= system::ONEFRAME_TIME * 255.f * 2.f;
-    color_.a(alpha_);
+	color_.a(static_cast<unsigned int>(alpha_));
     if (alpha_ < 0){
       return false;
     }
